src/main/steps/tasklistresponse/proceedings/ProceedingPostController.ts
import autobind from 'autobind-decorator';
import type { Response } from 'express';

import { CosApiClient } from '../../../app/case/CosApiClient';
import { Respondent } from '../../../app/case/definition';
import { toApiFormat } from '../../../app/case/to-api-format';
import type { AppRequest } from '../../../app/controller/AppRequest';
import { AnyObject, PostController } from '../../../app/controller/PostController';
import { FormFields, FormFieldsFn } from '../../../app/form/Form';
import { RESPOND_TO_APPLICATION } from '../../urls';

import { setProceedingDetails } from './ProceedingDetailsMapper';

@autobind
export class ProceedingPostController extends PostController<AnyObject> {
  constructor(protected readonly fields: FormFields | FormFieldsFn) {
    super(fields);
  }
  public async post(req: AppRequest, res: Response): Promise<void> {
    try {
      const caseworkerUser = req.session.user;
      const caseReference = req.session.userCase.id;

      const client = new CosApiClient(caseworkerUser.accessToken, 'https://return-url');

      const caseDataFromCos = await client.retrieveByCaseId(caseReference, caseworkerUser);
      Object.assign(req.session.userCase, caseDataFromCos);

      req.session.userCase?.respondents?.forEach((respondent: Respondent) => {
        if (respondent?.value?.user?.idamId === req.session?.user.id) {
          setProceedingDetails(req.session.userCase, respondent, req);
        }
      });
      const caseData = toApiFormat(req?.session?.userCase);
      caseData.id = caseReference;
      const updatedCaseDataFromCos = await client.updateCase(
        caseworkerUser,
        caseReference,
        caseData,
        'citizen-case-update'
      );
      Object.assign(req.session.userCase, updatedCaseDataFromCos);
      req.session.save(() => res.redirect(RESPOND_TO_APPLICATION));
    } catch (err) {
      throw new Error('ProceedingPostController - Case could not be updated.');
    }
  }
}
